requirements.txt
flask
flask-wtf
flask-ckeditor
python-dotenv
beautifulsoup4
docxtpl
flask-login
flask-wtf
openai
markdown
bleach
gunicorn
flask_sqlalchemy
tiktoken
flask-migrate
apscheduler
email-validator
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz
flask_limiter
flask-bcrypt
pytest
celery
redis
requests
beautifulsoup4
PyPDF2
python-dotenv
mistralai
openai
Flask
Flask-Session
openai-agents
